requirements.txt
fastapi==0.104.0
kafka==1.3.5
pydantic_settings==2.0.3
pyspark==3.5.0
python-decouple==3.8
Requests==2.31.0
schedule==1.2.1
